refactor: create exception_handler.py to centralize exception handling, removing logic from Users.py
